Changed scripts to extract data straight out of Hydstra.
Previous version used the Hydro database for extracting site info for
flow, water level, and precipitation. Because there are some
inconsistencies with the Hydro database it was decided to extract the
site info direct from Hydstra.

The script get_hydstra_sites.py results in a site summary for every site
in hydstra and writes it to a csv-file.

The script SW_quantity_site_info.py also adds river and aquifer
abstraction sites to the dataframe, does some filtering/organising, and
labels sites to primary, secondary, or other.